+++ src/core/payload/EventPayloadAttributes.ts
@@ -2,14 +2,11 @@
 
 export const TIMESTAMP = 'timestamp';
 export const EVENT_KIND = 'event.kind';
-export const DT_AGENT_VERSION = 'dt.agent.version';
-export const DT_AGENT_TECHNOLOGY_TYPE = 'dt.agent.technology_type';
-export const DT_AGENT_FLAVOR = 'dt.agent.flavor';
 export const APP_VERSION = 'app.version';
 export const OS_NAME = 'os.name';
 export const DEVICE_MANUFACTURER = 'device.manufacturer';
 export const DEVICE_MODEL_IDENTIFIER = 'device.model.identifier';
-export const WINDOW_ORIENTATION = 'window.orientation';
+export const EVENT_PROVIDER = 'event.provider';
 
 export const EVENT_KIND_RUM = 'RUM_EVENT';
 export const EVENT_KIND_BIZ = 'BIZ_EVENT';