+++ test/core/payload/EventPayload.spec.ts
@@ -10,7 +10,6 @@ import { EventPayload } from '../../../src/core/payload/EventPayload';
 import { openKitVersion } from '../../../src/core/PlatformConstants';
 import { DefaultRandomNumberProvider } from '../../../src/core/provider/DefaultRandomNumberProvider';
 import { TimestampProvider } from '../../../src/core/provider/TimestampProvider';
-import { SEND_TIMESTAMP_PLACEHOLDER } from '../../../src/core/utils/EventPayloadUtils';
 
 describe('EventPayload', () => {
     let config: Configuration;
@@ -70,24 +69,18 @@ describe('EventPayload', () => {
 
             expect(payloadJson['event.name']).toEqual('customName');
             expect(payloadJson.timestamp).toEqual(7000000);
-            expect(payloadJson['dt.application_id']).toEqual('application-id');
-            expect(payloadJson['event.kind']).toEqual('RUM_EVENT');
-            expect(payloadJson['dt.send_timestamp']).toEqual(
-                SEND_TIMESTAMP_PLACEHOLDER,
+            expect(payloadJson['dt.rum.application.id']).toEqual(
+                'application-id',
             );
-            expect(payloadJson['dt.instance_id']).toEqual('42');
-            expect(payloadJson['dt.sid']).toEqual('1234');
-            expect(payloadJson['dt.agent.version']).toEqual(openKitVersion);
-            expect(payloadJson['dt.agent.technology_type']).toEqual('openkit');
-            expect(payloadJson['dt.agent.flavor']).toEqual('nodejs');
+            expect(payloadJson['event.kind']).toEqual('RUM_EVENT');
+            expect(payloadJson['dt.rum.instance.id']).toEqual('42');
+            expect(payloadJson['dt.rum.sid']).toEqual('1234');
+            expect(payloadJson['event.provider']).toEqual('application-id');
 
             expect(payloadJson['app.version']).toEqual('1.0');
             expect(payloadJson['os.name']).toEqual('dynaOS');
             expect(payloadJson['device.manufacturer']).toEqual('dynatrace');
             expect(payloadJson['device.model.identifier']).toEqual('dynaPhone');
-            expect(payloadJson['window.orientation']).toEqual(
-                Orientation.Landscape,
-            );
         });
 
         it('should provide only enrichment data which is available', () => {
@@ -129,29 +122,26 @@ describe('EventPayload', () => {
 
             expect(payloadJson['event.name']).toEqual('customName');
             expect(payloadJson.timestamp).toEqual(7000000);
-            expect(payloadJson['dt.application_id']).toEqual('application-id');
-            expect(payloadJson['event.kind']).toEqual('RUM_EVENT');
-            expect(payloadJson['dt.send_timestamp']).toEqual(
-                SEND_TIMESTAMP_PLACEHOLDER,
+            expect(payloadJson['dt.rum.application.id']).toEqual(
+                'application-id',
             );
-            expect(payloadJson['dt.instance_id']).toEqual('42');
-            expect(payloadJson['dt.sid']).toEqual('1234');
-            expect(payloadJson['dt.agent.version']).toEqual(openKitVersion);
-            expect(payloadJson['dt.agent.technology_type']).toEqual('openkit');
-            expect(payloadJson['dt.agent.flavor']).toEqual('nodejs');
+            expect(payloadJson['event.kind']).toEqual('RUM_EVENT');
+            expect(payloadJson['dt.rum.instance.id']).toEqual('42');
+            expect(payloadJson['dt.rum.sid']).toEqual('1234');
+            expect(payloadJson['dt.rum.schema_version']).toEqual('1.0');
+            expect(payloadJson['event.provider']).toEqual('application-id');
 
             // Not available as they are not provided in the config
             expect(payloadJson['app.version']).toEqual(undefined);
             expect(payloadJson['os.name']).toEqual(undefined);
             expect(payloadJson['device.manufacturer']).toEqual(undefined);
             expect(payloadJson['device.model.identifier']).toEqual(undefined);
-            expect(payloadJson['window.orientation']).toEqual(undefined);
         });
 
-        it('should not be possible to override dt.sid', () => {
+        it('should not be possible to override dt.rum.sid', () => {
             // given
             const predefinedAttributes = {
-                'dt.sid': 'overridden',
+                'dt.rum.sid': 'overridden',
             };
 
             // then
@@ -162,14 +152,13 @@ describe('EventPayload', () => {
             );
             const payloadJson = JSON.parse(payload);
 
-            expect(payloadJson['dt.sid']).toEqual('1234');
-            expect(payloadJson['dt.overridden_keys']).toEqual(undefined);
+            expect(payloadJson['dt.rum.sid']).toEqual('1234');
         });
 
-        it('should not be possible to override dt.instance_id', () => {
+        it('should not be possible to override dt.rum.instance.id', () => {
             // given
             const predefinedAttributes = {
-                'dt.instance_id': 'overridden',
+                'dt.rum.instance.id': 'overridden',
             };
 
             // then
@@ -180,14 +169,13 @@ describe('EventPayload', () => {
             );
             const payloadJson = JSON.parse(payload);
 
-            expect(payloadJson['dt.instance_id']).toEqual('42');
-            expect(payloadJson['dt.overridden_keys']).toEqual(undefined);
+            expect(payloadJson['dt.rum.instance.id']).toEqual('42');
         });
 
-        it('should not be possible to override dt.send_timestamp', () => {
+        it('should not be possible to override dt.rum.application.id', () => {
             // given
             const predefinedAttributes = {
-                'dt.send_timestamp': 'overridden',
+                'dt.rum.application.id': 'overridden',
             };
 
             // then
@@ -198,16 +186,15 @@ describe('EventPayload', () => {
             );
             const payloadJson = JSON.parse(payload);
 
-            expect(payloadJson['dt.send_timestamp']).toEqual(
-                SEND_TIMESTAMP_PLACEHOLDER,
+            expect(payloadJson['dt.rum.application.id']).toEqual(
+                'application-id',
             );
-            expect(payloadJson['dt.overridden_keys']).toEqual(undefined);
         });
 
-        it('should not be possible to override dt.application_id', () => {
+        it('should not be possible to override dt.rum.schema_version', () => {
             // given
             const predefinedAttributes = {
-                'dt.application_id': 'overridden',
+                'dt.rum.schema_version': 'overridden',
             };
 
             // then
@@ -218,8 +205,24 @@ describe('EventPayload', () => {
             );
             const payloadJson = JSON.parse(payload);
 
-            expect(payloadJson['dt.application_id']).toEqual('application-id');
-            expect(payloadJson['dt.overridden_keys']).toEqual(undefined);
+            expect(payloadJson['dt.rum.schema_version']).toEqual('1.0');
+        });
+
+        it('should not be possible to override event.provider', () => {
+            // given
+            const predefinedAttributes = {
+                'event.provider': 'overridden',
+            };
+
+            // then
+            const payload = eventPayload.getCustomEventsPayload(
+                'customName',
+                predefinedAttributes,
+                1234,
+            );
+            const payloadJson = JSON.parse(payload);
+
+            expect(payloadJson['event.provider']).toEqual('application-id');
         });
 
         it('should be possible to override timestamp', () => {
@@ -237,7 +240,6 @@ describe('EventPayload', () => {
             const payloadJson = JSON.parse(payload);
 
             expect(payloadJson.timestamp).toEqual('overridden');
-            expect(payloadJson['dt.overridden_keys']).toEqual(['timestamp']);
         });
 
         it('should be possible to override dt.type', () => {
@@ -255,89 +257,6 @@ describe('EventPayload', () => {
             const payloadJson = JSON.parse(payload);
 
             expect(payloadJson['event.kind']).toEqual('overridden');
-            expect(payloadJson['dt.overridden_keys']).toEqual(['event.kind']);
-        });
-
-        it('should be possible to override dt.agent.version', () => {
-            // given
-            const predefinedAttributes = {
-                'dt.agent.version': 'overridden',
-            };
-
-            // then
-            const payload = eventPayload.getCustomEventsPayload(
-                'customName',
-                predefinedAttributes,
-                1234,
-            );
-            const payloadJson = JSON.parse(payload);
-
-            expect(payloadJson['dt.agent.version']).toEqual('overridden');
-            expect(payloadJson['dt.overridden_keys']).toEqual([
-                'dt.agent.version',
-            ]);
-        });
-
-        it('should be possible to override dt.agent.technology_type', () => {
-            // given
-            const predefinedAttributes = {
-                'dt.agent.technology_type': 'overridden',
-            };
-
-            // then
-            const payload = eventPayload.getCustomEventsPayload(
-                'customName',
-                predefinedAttributes,
-                1234,
-            );
-            const payloadJson = JSON.parse(payload);
-
-            expect(payloadJson['dt.agent.technology_type']).toEqual(
-                'overridden',
-            );
-            expect(payloadJson['dt.overridden_keys']).toEqual([
-                'dt.agent.technology_type',
-            ]);
-        });
-
-        it('should be possible to override dt.agent.flavor', () => {
-            // given
-            const predefinedAttributes = {
-                'dt.agent.flavor': 'overridden',
-            };
-
-            // then
-            const payload = eventPayload.getCustomEventsPayload(
-                'customName',
-                predefinedAttributes,
-                1234,
-            );
-            const payloadJson = JSON.parse(payload);
-
-            expect(payloadJson['dt.agent.flavor']).toEqual('overridden');
-            expect(payloadJson['dt.overridden_keys']).toEqual([
-                'dt.agent.flavor',
-            ]);
-        });
-
-        it('should be possible to override window.orientation', () => {
-            // given
-            const predefinedAttributes = {
-                'window.orientation': 'overridden',
-            };
-
-            // then
-            const payload = eventPayload.getCustomEventsPayload(
-                'customName',
-                predefinedAttributes,
-                1234,
-            );
-            const payloadJson = JSON.parse(payload);
-
-            expect(payloadJson['window.orientation']).toEqual('overridden');
-            expect(payloadJson['dt.overridden_keys']).toEqual([
-                'window.orientation',
-            ]);
         });
 
         it('should be possible to override os.name', () => {
@@ -355,7 +274,6 @@ describe('EventPayload', () => {
             const payloadJson = JSON.parse(payload);
 
             expect(payloadJson['os.name']).toEqual('overridden');
-            expect(payloadJson['dt.overridden_keys']).toEqual(['os.name']);
         });
 
         it('should be possible to override device.manufacturer', () => {
@@ -373,9 +291,6 @@ describe('EventPayload', () => {
             const payloadJson = JSON.parse(payload);
 
             expect(payloadJson['device.manufacturer']).toEqual('overridden');
-            expect(payloadJson['dt.overridden_keys']).toEqual([
-                'device.manufacturer',
-            ]);
         });
 
         it('should be possible to override app.version', () => {
@@ -393,7 +308,6 @@ describe('EventPayload', () => {
             const payloadJson = JSON.parse(payload);
 
             expect(payloadJson['app.version']).toEqual('2.0');
-            expect(payloadJson['dt.overridden_keys']).toEqual(['app.version']);
         });
 
         it('should be possible to override device.model.identifier', () => {
@@ -413,9 +327,6 @@ describe('EventPayload', () => {
             expect(payloadJson['device.model.identifier']).toEqual(
                 'overridden',
             );
-            expect(payloadJson['dt.overridden_keys']).toEqual([
-                'device.model.identifier',
-            ]);
         });
 
         it('should override the name', () => {
@@ -438,7 +349,7 @@ describe('EventPayload', () => {
             // given
             const predefinedAttributes = {
                 'dt.test': 'test',
-                'dt.application_id': 'myID',
+                'dt.rum.application.id': 'myID',
                 'dt': {
                     test: 'test',
                 },
@@ -453,7 +364,7 @@ describe('EventPayload', () => {
             expect(payload).not.toContain('dt.test');
             expect(payload).not.toContain('dt:');
             expect(payload).not.toContain('myID');
-            expect(payload).toContain('dt.application_id');
+            expect(payload).toContain('dt.rum.application.id');
             expect(payload).toContain('customName');
         });
     });
@@ -476,24 +387,19 @@ describe('EventPayload', () => {
             expect(payloadJson['event.name']).toEqual('predefined');
             expect(payloadJson['event.type']).toEqual('customType');
             expect(payloadJson.timestamp).toEqual(7000000);
-            expect(payloadJson['dt.application_id']).toEqual('application-id');
-            expect(payloadJson['event.kind']).toEqual('BIZ_EVENT');
-            expect(payloadJson['dt.send_timestamp']).toEqual(
-                SEND_TIMESTAMP_PLACEHOLDER,
+            expect(payloadJson['dt.rum.application.id']).toEqual(
+                'application-id',
             );
-            expect(payloadJson['dt.instance_id']).toEqual('42');
-            expect(payloadJson['dt.sid']).toEqual('1234');
-            expect(payloadJson['dt.agent.version']).toEqual(openKitVersion);
-            expect(payloadJson['dt.agent.technology_type']).toEqual('openkit');
-            expect(payloadJson['dt.agent.flavor']).toEqual('nodejs');
+            expect(payloadJson['event.kind']).toEqual('BIZ_EVENT');
+            expect(payloadJson['dt.rum.instance.id']).toEqual('42');
+            expect(payloadJson['dt.rum.sid']).toEqual('1234');
+            expect(payloadJson['dt.rum.schema_version']).toEqual('1.0');
+            expect(payloadJson['event.provider']).toEqual('application-id');
 
             expect(payloadJson['app.version']).toEqual('1.0');
             expect(payloadJson['os.name']).toEqual('dynaOS');
             expect(payloadJson['device.manufacturer']).toEqual('dynatrace');
             expect(payloadJson['device.model.identifier']).toEqual('dynaPhone');
-            expect(payloadJson['window.orientation']).toEqual(
-                Orientation.Landscape,
-            );
         });
 
         it('should provide only enrichment data which is available', () => {
@@ -536,23 +442,20 @@ describe('EventPayload', () => {
             expect(payloadJson['event.name']).toEqual('predefined');
             expect(payloadJson['event.type']).toEqual('customType');
             expect(payloadJson.timestamp).toEqual(7000000);
-            expect(payloadJson['dt.application_id']).toEqual('application-id');
-            expect(payloadJson['event.kind']).toEqual('BIZ_EVENT');
-            expect(payloadJson['dt.send_timestamp']).toEqual(
-                SEND_TIMESTAMP_PLACEHOLDER,
+            expect(payloadJson['dt.rum.application.id']).toEqual(
+                'application-id',
             );
-            expect(payloadJson['dt.instance_id']).toEqual('42');
-            expect(payloadJson['dt.sid']).toEqual('1234');
-            expect(payloadJson['dt.agent.version']).toEqual(openKitVersion);
-            expect(payloadJson['dt.agent.technology_type']).toEqual('openkit');
-            expect(payloadJson['dt.agent.flavor']).toEqual('nodejs');
+            expect(payloadJson['event.kind']).toEqual('BIZ_EVENT');
+            expect(payloadJson['dt.rum.instance.id']).toEqual('42');
+            expect(payloadJson['dt.rum.sid']).toEqual('1234');
+            expect(payloadJson['dt.rum.schema_version']).toEqual('1.0');
+            expect(payloadJson['event.provider']).toEqual('application-id');
 
             // Not available as they are not provided in the config
             expect(payloadJson['app.version']).toEqual(undefined);
             expect(payloadJson['os.name']).toEqual(undefined);
             expect(payloadJson['device.manufacturer']).toEqual(undefined);
             expect(payloadJson['device.model.identifier']).toEqual(undefined);
-            expect(payloadJson['window.orientation']).toEqual(undefined);
         });
 
         it('should take the type as name if name is not available', () => {
@@ -569,13 +472,12 @@ describe('EventPayload', () => {
 
             expect(payloadJson['event.name']).toEqual('customType');
             expect(payloadJson['event.type']).toEqual('customType');
-            expect(payloadJson['dt.overridden_keys']).toEqual(undefined);
         });
 
-        it('should not be possible to override dt.sid', () => {
+        it('should not be possible to override dt.rum.sid', () => {
             // given
             const predefinedAttributes = {
-                'dt.sid': 'overridden',
+                'dt.rum.sid': 'overridden',
             };
 
             // then
@@ -586,14 +488,13 @@ describe('EventPayload', () => {
             );
             const payloadJson = JSON.parse(payload);
 
-            expect(payloadJson['dt.sid']).toEqual('1234');
-            expect(payloadJson['dt.overridden_keys']).toEqual(undefined);
+            expect(payloadJson['dt.rum.sid']).toEqual('1234');
         });
 
-        it('should not be possible to override dt.instance_id', () => {
+        it('should not be possible to override dt.rum.instance.id', () => {
             // given
             const predefinedAttributes = {
-                'dt.instance_id': 'overridden',
+                'dt.rum.instance.id': 'overridden',
             };
 
             // then
@@ -604,14 +505,13 @@ describe('EventPayload', () => {
             );
             const payloadJson = JSON.parse(payload);
 
-            expect(payloadJson['dt.instance_id']).toEqual('42');
-            expect(payloadJson['dt.overridden_keys']).toEqual(undefined);
+            expect(payloadJson['dt.rum.instance.id']).toEqual('42');
         });
 
-        it('should not be possible to override dt.send_timestamp', () => {
+        it('should not be possible to override dt.rum.application.id', () => {
             // given
             const predefinedAttributes = {
-                'dt.send_timestamp': 'overridden',
+                'dt.rum.application.id': 'overridden',
             };
 
             // then
@@ -622,16 +522,15 @@ describe('EventPayload', () => {
             );
             const payloadJson = JSON.parse(payload);
 
-            expect(payloadJson['dt.send_timestamp']).toEqual(
-                SEND_TIMESTAMP_PLACEHOLDER,
+            expect(payloadJson['dt.rum.application.id']).toEqual(
+                'application-id',
             );
-            expect(payloadJson['dt.overridden_keys']).toEqual(undefined);
         });
 
-        it('should not be possible to override dt.application_id', () => {
+        it('should not be possible to override dt.rum.schema_version', () => {
             // given
             const predefinedAttributes = {
-                'dt.application_id': 'overridden',
+                'dt.rum.schema_version': 'overridden',
             };
 
             // then
@@ -642,8 +541,24 @@ describe('EventPayload', () => {
             );
             const payloadJson = JSON.parse(payload);
 
-            expect(payloadJson['dt.application_id']).toEqual('application-id');
-            expect(payloadJson['dt.overridden_keys']).toEqual(undefined);
+            expect(payloadJson['dt.rum.schema_version']).toEqual('1.0');
+        });
+
+        it('should be possible to override event.provider', () => {
+            // given
+            const predefinedAttributes = {
+                'event.provider': 'overridden',
+            };
+
+            // then
+            const payload = eventPayload.getBizEventsPayload(
+                'customType',
+                predefinedAttributes,
+                1234,
+            );
+            const payloadJson = JSON.parse(payload);
+
+            expect(payloadJson['event.provider']).toEqual('overridden');
         });
 
         it('should be possible to override timestamp', () => {
@@ -661,7 +576,6 @@ describe('EventPayload', () => {
             const payloadJson = JSON.parse(payload);
 
             expect(payloadJson.timestamp).toEqual('overridden');
-            expect(payloadJson['dt.overridden_keys']).toEqual(['timestamp']);
         });
 
         it('should not be possible to override dt.type', () => {
@@ -679,89 +593,6 @@ describe('EventPayload', () => {
             const payloadJson = JSON.parse(payload);
 
             expect(payloadJson['event.kind']).toEqual('BIZ_EVENT');
-            expect(payloadJson['dt.overridden_keys']).toEqual(undefined);
-        });
-
-        it('should be possible to override dt.agent.version', () => {
-            // given
-            const predefinedAttributes = {
-                'dt.agent.version': 'overridden',
-            };
-
-            // then
-            const payload = eventPayload.getBizEventsPayload(
-                'customType',
-                predefinedAttributes,
-                1234,
-            );
-            const payloadJson = JSON.parse(payload);
-
-            expect(payloadJson['dt.agent.version']).toEqual('overridden');
-            expect(payloadJson['dt.overridden_keys']).toEqual([
-                'dt.agent.version',
-            ]);
-        });
-
-        it('should be possible to override dt.agent.technology_type', () => {
-            // given
-            const predefinedAttributes = {
-                'dt.agent.technology_type': 'overridden',
-            };
-
-            // then
-            const payload = eventPayload.getBizEventsPayload(
-                'customType',
-                predefinedAttributes,
-                1234,
-            );
-            const payloadJson = JSON.parse(payload);
-
-            expect(payloadJson['dt.agent.technology_type']).toEqual(
-                'overridden',
-            );
-            expect(payloadJson['dt.overridden_keys']).toEqual([
-                'dt.agent.technology_type',
-            ]);
-        });
-
-        it('should be possible to override dt.agent.flavor', () => {
-            // given
-            const predefinedAttributes = {
-                'dt.agent.flavor': 'overridden',
-            };
-
-            // then
-            const payload = eventPayload.getBizEventsPayload(
-                'customType',
-                predefinedAttributes,
-                1234,
-            );
-            const payloadJson = JSON.parse(payload);
-
-            expect(payloadJson['dt.agent.flavor']).toEqual('overridden');
-            expect(payloadJson['dt.overridden_keys']).toEqual([
-                'dt.agent.flavor',
-            ]);
-        });
-
-        it('should be possible to override window.orientation', () => {
-            // given
-            const predefinedAttributes = {
-                'window.orientation': 'overridden',
-            };
-
-            // then
-            const payload = eventPayload.getBizEventsPayload(
-                'customType',
-                predefinedAttributes,
-                1234,
-            );
-            const payloadJson = JSON.parse(payload);
-
-            expect(payloadJson['window.orientation']).toEqual('overridden');
-            expect(payloadJson['dt.overridden_keys']).toEqual([
-                'window.orientation',
-            ]);
         });
 
         it('should be possible to override os.name', () => {
@@ -779,7 +610,6 @@ describe('EventPayload', () => {
             const payloadJson = JSON.parse(payload);
 
             expect(payloadJson['os.name']).toEqual('overridden');
-            expect(payloadJson['dt.overridden_keys']).toEqual(['os.name']);
         });
 
         it('should be possible to override device.manufacturer', () => {
@@ -797,9 +627,6 @@ describe('EventPayload', () => {
             const payloadJson = JSON.parse(payload);
 
             expect(payloadJson['device.manufacturer']).toEqual('overridden');
-            expect(payloadJson['dt.overridden_keys']).toEqual([
-                'device.manufacturer',
-            ]);
         });
 
         it('should be possible to override app.version', () => {
@@ -817,7 +644,6 @@ describe('EventPayload', () => {
             const payloadJson = JSON.parse(payload);
 
             expect(payloadJson['app.version']).toEqual('2.0');
-            expect(payloadJson['dt.overridden_keys']).toEqual(['app.version']);
         });
 
         it('should be possible to override device.model.identifier', () => {
@@ -837,9 +663,6 @@ describe('EventPayload', () => {
             expect(payloadJson['device.model.identifier']).toEqual(
                 'overridden',
             );
-            expect(payloadJson['dt.overridden_keys']).toEqual([
-                'device.model.identifier',
-            ]);
         });
 
         it('should override the name', () => {
@@ -862,7 +685,7 @@ describe('EventPayload', () => {
             // given
             const predefinedAttributes = {
                 'dt.test': 'test',
-                'dt.application_id': 'myID',
+                'dt.rum.application.id': 'myID',
                 'dt': {
                     test: 'test',
                 },
@@ -877,7 +700,7 @@ describe('EventPayload', () => {
             expect(payload).not.toContain('dt.test');
             expect(payload).not.toContain('dt:');
             expect(payload).not.toContain('myID');
-            expect(payload).toContain('dt.application_id');
+            expect(payload).toContain('dt.rum.application.id');
             expect(payload).toContain('customType');
         });
 
@@ -894,7 +717,7 @@ describe('EventPayload', () => {
                 1234,
             );
 
-            expect(payload).toContain('dt.application_id');
+            expect(payload).toContain('dt.rum.application.id');
             expect(payload).toContain('customType');
         });
     });